+++ engine/lumen-host/src/platform/windows/media_queue.rs
@@ -113,7 +113,7 @@ mod tests {
         PlatformEncodedVideoFrame {
             payload: vec![value],
             decoder_configuration_record: None,
-            presentation_time_90khz: u32::from(value),
+            presentation_time_90khz: u64::from(value),
             key_frame,
             requires_bootstrap_acknowledgement: false,
             repair_keyframe: false,
